+++ R/db.R
@@ -222,4 +222,3 @@ unique_path <- function(path) {
         a
     }
 }
-