+++ R/parse.R
@@ -162,4 +162,3 @@ parse_dcf_list <- function(x) {
     parts <- parts[nchar(parts) > 0L]
     parts
 }
-